Prevent timer updates from double-starting timers
`self:GetTime(...) == 0` is not correct check for a timer that doesn't
exist yet, because `GetTime` returns zero if timer was just created and
hasn't ticked yet. `not self:IsStarted(...)` would work, but since we
format the id string and call `DBT:GetBar()` anyway we can avoid
duplicating that work by just checking `not bar`.