Add support for loading mods by MapID
X-DBM-Mod-MapID has a bit of a confusing name, it actually expects you to pass instance IDs. That is not particularly helpful for mods to be loaded in specific outdoor areas since whole continents are instances.

Add feature to pass an id prefixed with m to actually handle it as map ID. Same logic for getting the mod name from an m-prefixed mod id like it works for z-prefixed ones.