Fix some of the bigger problems for LuaLS.
This unlocks amazing LuaLS features like auto-complete and
jump-to-definition for all fields in DBM, mods, timers, warnings etc :)

The main problem was that the DBM the DBM global was registered in a way
it didn't understand (and hence it only saw about half the fields).
Also, we had some cases of functions being tagged as optional for no
good reason triggering a lot of nil check warnings.

It currently still reports around 600 problems or so, but a lot of these
are just it being unhappy about us setting random fields in Frame
objects which I guess we should disable.