Add LuaLS annotations to DBM-GUI
Frames are a bit messy because they are classes, so if you want a frame
with a custom field in it you need to declare a new subclass. Also,
frames using templates have to declare a new class inheriting from the
frame and template(s). That's why the diff is relatively large.