Bump versions again and make it a mandatory (force disable) triggering update do to how important bug fix is since the function that was broken is used A LOT by a lot of mods and core functions, especially affecting wrath classic but also retail and classic era in certain situations custom "boss unit Id only" checks were used.
Long short of it. In this weeks updates, a function was rewritten to be better but had a bug in it that broke the rewrite in specific circomstances that were missed in testing because I didn't do the testing on wrath classic (the only client that has yet to recieve UnitTokenFromGUID from blizzard.