Create LuaLS workflow
Example of this in action: https://github.com/emmericp/test-gh-actions/pull/2/files#annotation_18742515697

This will not block a pull request or release, it just adds warnings as
annotations to files. It will report success even if a warning was reported.
Annotations only show up on files you are touching in the PR, but you can
always view the log to see all warnings.

This triggers a few warnings which are fixed in https://github.com/DeadlyBossMods/DBM-Unified/pull/446